Fix sprites distance int cast

+++ includes/canvas.h
@@ -6,7 +6,7 @@
 /*   By: chamada <chamada@student.le-101.fr>        +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2019/12/05 15:05:34 by chamada      #+#   ##    ##    #+#       */
-/*   Updated: 2020/01/12 02:30:43 by chamada     ###    #+. /#+    ###.fr     */
+/*   Updated: 2020/01/19 01:41:57 by chamada     ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
@@ -17,6 +17,6 @@
 # include <environment.h>
 
 int		init_canvas(t_env *env);
-void	draw_tex(t_env *env, int pos_x, int z, int size);
+void	draw_tex(t_env *env, int pos_x, float z, int size);
 
 #endif
